Added Python Solution
This is my hello world python script as it is my first time writing in
python.

I like it

I've decided to implement this with a slightly different algorithm that
loops over loading data, parsing and processing it with no storage of
all data before processins.

It's an easy read